ISSUE #5312 - refactor
